Add Camera tweening to desk animation
- Add animation to tween the Camera from the room to the desk when
clicking the portfolio button in the Hero component
- Expose functions to manipulate the camera to the Scene component, so
that can be called from other components (eg: Hero component when
clicking the portfolio button)
- Add TweenJS package
- Refactor the canvas style. To make the room appear on the right side,
the canvas size was reduced to ~40% and moved right. This worked fine
but made it much harder to animate the camera to the desk since we had
to expand the canvas to the full width of the screen, which caused the
viewport aspect ratio to be off. To fix this, the canvas is always 100%
width, and we simply move the viewport right which has the same effect
but is much easier to work with
- Clean up the Floor component
- Use CSS variables to set the background colors, for both the HTML and
the scene