Refactor skills list props
Refactor the skills list to take an array of skills and associate the
skill to its name and icon in the SkillsList component directly rather
than importing the skills in each file where we want to create a list.
This makes the code more readable and allows for easier maintenance.