Fix animation not stopping when opening portfolio
This commit fixes an issue where the room animation would not stop when
the user clicked the view portfolio button. This has been fixed by
forcing the scale of the objects based on the state of the animation
(started, running, stopped). For objects which position is tweened, the
tween is fast forwarded when the user opens the portfolio to move them
to their final position.